requirements.txt
Django==5.1.7
djangorestframework==3.15.2
pytest==8.3.5
pytest-cov==6.0.0
pytest-django==4.10.0